+++ test/cases/javascript-assets-compile/src/index.js
@@ -1,10 +1,13 @@
-const tmpl = require('Template/widget.pug');
-
-console.log(tmpl({
+const locals = {
   data: {
     items: {
       a: 123,
-      b: 'abc'
-    }
-  }
-}));
+      b: 'abc',
+    },
+  },
+};
+
+const tmpl = require('Templates/widget.pug');
+const html = tmpl(locals);
+
+console.log(html);