+++ test/cases/javascript-assets-render/src/index.js
@@ -1,2 +1,4 @@
-const html = require('Template/widget.pug');
+// pass data into pug with `render` method
+const html = require('Templates/widget.pug?{"data":{"items":{"a":123,"b":"abc"}}}');
+
 console.log(html);